final-projects/backend/BillingService/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.0.3
reportlab==4.0.7
jinja2==3.1.2
python-multipart==0.0.6
requests==2.31.0
python-dateutil==2.8.2
babel==2.13.1 